+++ src/request/email.ts
@@ -322,6 +322,7 @@ export default class Email implements EmailProps {
     let domain = address.substring(atIndex + 1);
 
     domain = this.cleanDomain(domain);
+    localPart = localPart.normalize('NFC');
 
     let separator = '+';
     if (Object.prototype.hasOwnProperty.call(Email.yahooDomains, domain)) {
@@ -363,7 +364,7 @@ export default class Email implements EmailProps {
     // We don't need to strip a trailing '.' because validation (isEmail())
     // rejects domains that have it.
 
-    domain = punycode.toASCII(domain);
+    domain = punycode.toASCII(domain.normalize('NFC'));
 
     domain = domain.replace(/(?:\.com){2,}$/, '.com');
     domain = domain.replace(/^\d+(?:gmail?\.com)$/, 'gmail.com');