+++ CHANGELOG.md
@@ -27,6 +27,10 @@ CHANGELOG
   become `gmail.com`.
 * Additional `gmail.com` typos are now normalized when `hashAddress` is
   used. For example, `gmali.com` will become `gmail.com`.
+* When `hashAddress` is used, the domain part of an email address is now
+  normalized to NFC.
+* When `hashAddress` is used, the local part of an email address is now
+  normalized to NFC.
 
 6.0.0 (2023-12-05)
 ------------------